+++ src/oemof/network/groupings.py
@@ -20,7 +20,7 @@ from collections.abc import MutableMapping
 from itertools import chain
 from itertools import filterfalse
 
-from oemof.network.network import Edge
+from oemof.network.network.edge import Edge
 
 # TODO: Update docstrings.
 #
@@ -122,7 +122,7 @@ class Grouping:
                 + "one of `key` or `constant_key`."
             )
         self.filter = filter
-        for kw in ["value", "merge", "filter"]:
+        for kw in ["value", "merge"]:
             if kw in kwargs:
                 setattr(self, kw, kwargs[kw])